refactor: batch 아래 job 패키지 삭제
